feat: add CRUD endpoints for TodoController (#3)
This commit introduces CRUD (Create, Read, Update, Delete) endpoints for the TodoController:

- **Create**: Added `Create` endpoint for creating new todos.
- **Delete**: Added `Delete` endpoint for deleting todos by ID.
- **Get All**: Added `GetAll` endpoint for retrieving all todos with pagination.
- **Get by ID**: Added `GetById` endpoint for retrieving a todo by its ID.
- **Get by Title**: Added `GetByTitle` endpoint for retrieving a todo by its title.
- **Update**: Added `Update` endpoint for updating existing todos.
- **Update Status**: Added `UpdateStatus` endpoint for updating the status of a todo.

Each endpoint uses MediatR for handling requests and responses, ensuring separation of concerns and maintainability.

+++ Core/Application/Features/Todos/UpdateTodoStatus/UpdateTodoStatusCommandValidator.cs
@@ -16,7 +16,7 @@ public class UpdateTodoStatusCommandValidator : AbstractValidator<UpdateTodoStat
         RuleFor(todo => todo.Status)
          .NotEmpty().WithMessage("{PropertyName} is required.")
          .NotNull()
-         .Must(x => Enum.IsDefined(typeof(TodoStatusEnum), x))
+         .Must(x => Enum.IsDefined(typeof(TodoStatusEnum), (int)x))
          .WithMessage("{PropertyName} is not valid.");
     }
 }
